requirements.txt
# Core RL libraries
gymnasium>=0.29.0
stable-baselines3>=2.0.0
ray[rllib]>=2.8.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# UI and interfaces
streamlit>=1.25.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
tqdm>=4.65.0
rich>=13.0.0
